pyproject.toml
[project]
name = "zecathon-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.121.1",
    "jinja2>=3.1.6",
    "pdf417gen>=0.8.1",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "python-barcode[images]>=0.16.1",
    "segno>=1.6.6",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "ruff>=0.14.4",
]
